+++ src/scripts/minify-json-files.js
